// webapp/src/api/model/talent.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PrimaryAttribute } from './primaryAttribute';

export interface Talent {
  firstAttribute: PrimaryAttribute;
  id?: string;
  name: string;
  secondAttribute: PrimaryAttribute;
  tags: string[];
  thirdAttribute: PrimaryAttribute;
}